+++ General/Misc.vb
@@ -7,7 +7,7 @@ Imports System.Management
 Imports System.Runtime.InteropServices
 Imports System.Text
 Imports System.Text.RegularExpressions
-
+Imports Microsoft.Win32
 Imports StaxRip.UI
 
 Public Module ShortcutModule
@@ -1519,15 +1519,10 @@ Public Class FileTypes
 End Class
 
 Public Class OSVersion
-    Shared Property Windows7 As Single = 6.1
-    Shared Property Windows8 As Single = 6.2
-    Shared Property Windows10 As Single = 10.0
-
-    Shared ReadOnly Property Current As Single
-        Get
-            Return CSng(Environment.OSVersion.Version.Major + Environment.OSVersion.Version.Minor / 10)
-        End Get
-    End Property
+    Shared ReadOnly Property Windows7 As Single = 6.1
+    Shared ReadOnly Property Windows8 As Single = 6.2
+    Shared ReadOnly Property Windows10 As Single = 10.0
+    Shared ReadOnly Property Current As Single = CSng(Environment.OSVersion.Version.Major + Environment.OSVersion.Version.Minor / 10)
 End Class
 
 Public Class OS
@@ -3855,3 +3850,38 @@ End Enum
 Public Interface IUpdateUI
     Sub UpdateUI()
 End Interface
+
+Public Class TextEncoding
+    Shared Property EncodingOfProcess As Encoding = Encoding.Default
+    Shared Property CodePageOfProcess As Integer = Encoding.Default.CodePage
+    Shared Property CodePageOfSystem As Integer = Registry.LocalMachine.GetString("SYSTEM\CurrentControlSet\Control\Nls\CodePage", "ACP").ToInt
+    Shared Property EncodingOfSystem As Encoding = Encoding.GetEncoding(CodePageOfSystem)
+    Shared Property PathsSupportedBySystemEncoding As New Dictionary(Of String, Boolean)
+    Shared Property UTF8CodePage As Integer = 65001
+
+    Shared Function IsPathSupportedBySystemEncoding(path As String) As Boolean
+        If path = "" Then
+            Return True
+        End If
+
+        If Not PathsSupportedBySystemEncoding.ContainsKey(path) Then
+            PathsSupportedBySystemEncoding(path) = path.IsSystemEncodingCompatible
+        End If
+
+        Return PathsSupportedBySystemEncoding(path)
+    End Function
+
+    Shared Function IsSystemUTF8() As Boolean
+        Return CodePageOfSystem = UTF8CodePage
+    End Function
+
+    Shared Function AvsEncoderSupportsUTF8(Optional commandLine As String = Nothing) As Boolean
+        If commandLine = "" Then
+            commandLine = p.VideoEncoder.GetCommandLine(True, True)
+        End If
+
+        If commandLine <> "" Then
+            Return commandLine.Contains("x264") AndAlso commandLine.Contains(".avs") AndAlso Not g.ContainsPipeTool(commandLine)
+        End If
+    End Function
+End Class