+++ General/Muxer.vb
@@ -173,7 +173,7 @@ Public MustInherit Class Muxer
                 If v.Contains(VB6.ChrW(&HA) + VB6.ChrW(&H0) + VB6.ChrW(&HD) + VB6.ChrW(&HA)) Then
                     v = v.FixBreak
                     v = v.Replace(BR + VB6.ChrW(&H0) + BR, BR + "langidx: 0" + BR)
-                    File.WriteAllText(fp, v, Encoding.Default)
+                    v.WriteFileSystemEncoding(fp)
                 End If
             End If
 
@@ -486,26 +486,29 @@ Public Class BatchMuxer
     End Function
 
     Overrides Sub Mux()
-        Log.WriteHeader("Batch Muxing")
+        Log.WriteHeader("Command Line Muxing")
 
-        Dim batchPath = p.TempDir + p.TargetFile.Base + "_mux.bat"
-        Dim batchCode = Proc.WriteBatchFile(batchPath, Macro.Expand(CommandLines))
+        For Each line In Macro.Expand(CommandLines.Trim).SplitLinesNoEmpty
+            Using proc As New Proc
+                proc.Header = "Command Line Muxing: " + Name
 
-        Using proc As New Proc
-            proc.Header = "Video encoding command line encoder: " + Name
-            proc.WriteLog(batchCode + BR2)
-            proc.File = "cmd.exe"
-            proc.Arguments = "/C call """ + batchPath + """"
+                If line.Contains("|") Then
+                    proc.File = "cmd.exe"
+                    proc.Arguments = "/S /C """ + line + """"
+                Else
+                    proc.CommandLine = line
+                End If
 
-            Try
-                proc.Start()
-            Catch ex As AbortException
-                Throw ex
-            Catch ex As Exception
-                g.ShowException(ex)
-                Throw New AbortException
-            End Try
-        End Using
+                Try
+                    proc.Start()
+                Catch ex As AbortException
+                    Throw ex
+                Catch ex As Exception
+                    g.ShowException(ex)
+                    Throw New AbortException
+                End Try
+            End Using
+        Next
     End Sub
 
     Overrides Function Edit() As DialogResult