+++ General/Project.vb
@@ -233,13 +233,6 @@ Public Class Project
                     Exit Property
                 End If
 
-                If Text.Encoding.Default.CodePage <> 65001 AndAlso
-                    Not value.IsANSICompatible AndAlso Script.Engine = ScriptEngine.AviSynth Then
-
-                    MsgWarn(Strings.NoUnicode)
-                    Exit Property
-                End If
-
                 TargetFileValue = value
                 NotifyPropertyChanged()
             End If