+++ General/Strings.vb
@@ -1,6 +1,5 @@
 
 Public Class Strings
     Public Shared Muxer As String = "A muxer merges different video, audio and subtitle files into a single container file which is the actual output file."
-    Public Shared NoUnicode As String = $"Unicode filenames are not supported by AviSynth.{BR2}VapourSynth supports unicode, it can be enabled at:{BR2}Filters > Filter Setup > VapourSynth"
     Public Shared TaskDialogFooter As String = "[copymsg: Copy Message]"
 End Class