+++ My Project/AssemblyInfo.vb
@@ -7,6 +7,6 @@ Imports System.Runtime.InteropServices
 <Assembly: AssemblyCopyright("Copyright (C) 2002-2020 StaxRip authors")>
 <Assembly: ComVisible(False)> 
 <Assembly: Guid("54d5dc0f-6ba7-42f0-a82f-9c71d561cb8b")>
-<Assembly: AssemblyVersion("2.1.5.4")>
-<Assembly: AssemblyFileVersion("2.1.5.4")>
+<Assembly: AssemblyVersion("2.1.5.5")>
+<Assembly: AssemblyFileVersion("2.1.5.5")>
 <Assembly: AssemblyDescription("Video encoding app")>